Convert template lock to use async_track_template_result
